==> 4_NLP_classification.ipynb <==
import pandas as pd
import json
import urllib
import os
import random
from urllib.parse import quote
from IPython.display import Image
from datetime import datetime


paintings = pd.read_csv('data/wikidata_all_paintings.csv')
paintings = paintings.drop_duplicates(subset='item', keep='first')
food_keywords = pd.read_csv('labels/food_related_keywords.csv')

display(paintings.head())
display(food_keywords)

word_to_food_term = food_keywords.set_index('word')['closest_food_term'].str.lower().to_dict()

unique_food_terms = food_keywords['closest_food_term'].str.lower().unique()

paintings['depicts'] = paintings['depicts'].fillna('')

paintings['depicts_list'] = paintings['depicts'].str.lower().str.split(',').apply(lambda x: [word.strip() for word in x])

depicts_exploded = paintings.explode('depicts_list')

depicts_exploded['closest_food_term'] = depicts_exploded['depicts_list'].map(word_to_food_term)

matched_depicts = depicts_exploded.dropna(subset=['closest_food_term'])

food_dummies = pd.get_dummies(matched_depicts['closest_food_term'])

food_dummies = food_dummies.reindex(columns=unique_food_terms, fill_value=0)

food_dummies_grouped = food_dummies.groupby(matched_depicts.index).max()

paintings = paintings.join(food_dummies_grouped)

paintings[unique_food_terms] = paintings[unique_food_terms].fillna(0).astype(int)

for term in unique_food_terms:
    if term not in paintings.columns:
        paintings[term] = 0

paintings = paintings.drop(columns=['depicts_list'])

display(paintings.head())


columns_to_sum = unique_food_terms

paintings['food_count'] = paintings[columns_to_sum].sum(axis=1)

paintings = paintings.sort_values('food_count', ascending=False)

paintings.dropna(subset=['image_url'], inplace=True)


def get_image_path(url,path='img/img_512/'):
    parsed = urllib.parse.urlparse(url)
    first_decode = urllib.parse.unquote(parsed.path)
    filename = os.path.basename(first_decode)
    return path + filename

paintings['image_path'] = paintings['image_url'].apply(get_image_path)

paintings_with_food = paintings[['item','image_path']].join(paintings[unique_food_terms])


paintings_with_food = paintings_with_food[~paintings_with_food['image_path'].str.endswith('.tif')]
paintings_with_food = paintings_with_food[paintings_with_food['image_path'].apply(lambda x: os.path.exists(x))]


display(paintings_with_food)

paintings_with_food.to_csv('data/paintings_with_food_nlp.csv', index=False)

paintings_image_path = paintings_with_food[['image_path','item']]
paintings_image_path.to_csv('data/paintings_image_path.csv', index=False)

# Filter paintings that have at least one food-related term
paintings_subset = paintings_with_food[paintings_with_food[unique_food_terms].sum(axis=1) > 0]
paintings_subset


paintings_subset = paintings_subset.drop(columns='item')
paintings_subset = paintings_subset.reset_index(drop=True)
paintings_subset.to_csv('data/paintings_subset.csv', index=False)

# Define the label columns
label_columns = unique_food_terms

# Function to create predictions
def create_prediction(row):
    selected_labels = [label.capitalize() for label in label_columns if row[label] == 1]
    if selected_labels:
        return {
            "result": [
                {
                    "from_name": "labels",
                    "to_name": "image",
                    "type": "choices",
                    "value": {
                        "choices": selected_labels
                    }
                }
            ]
        }
    else:
        return {"result": []}

# Create the Label Studio JSON structure with predictions
label_studio_data = []
for _, row in paintings_with_food.iterrows():
    data_entry = {
        "data": {
            "image": '/data/local-files/?d='+row['image_path'] # Ensure the path is correct
        },
        "predictions": [create_prediction(row)]
    }
    label_studio_data.append(data_entry)


timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')

# Save to JSON file with timestamp in the filename
with open(f'labels/label_studio_predictions_{timestamp}.json', 'w') as f:
    json.dump(label_studio_data, f, indent=4)
